Assay schedule data types
